+++ src/main/kotlin/no/ssb/timeusesurveyapi/companion/CompanionGateway.kt
@@ -22,3 +22,7 @@ class CompanionGateway(
     internal fun deleteCompanions(respondentId: UUID, sessionToken: SessionToken) =
         service.makeRequest(RequestWrapper(DELETE, deleteCompanionPath(respondentId), sessionToken))
 }
+
+internal fun getCompanionPath(respondentId: UUID) = "/v1/hangingwith/$respondentId"
+internal fun getCompanionByIdPath(respondentId: UUID, id: String) = "/v1/hangingwith/$respondentId/$id"
+internal fun deleteCompanionPath(respondentId: UUID) = "/v1/hangingwith/$respondentId"
