+++ src/main/kotlin/no/ssb/timeusesurveyapi/diarystarthistory/DiaryStartHistoryGateway.kt
@@ -16,12 +16,19 @@ class DiaryStartHistoryGateway(
     internal fun getDiaryStartHistories(respondentId: UUID, sessionToken: SessionToken) =
         service.makeRequest(RequestWrapper(GET, getDiaryStartHistoriesPath(respondentId), sessionToken))
 
-    internal fun postDiaryStartHistory(respondentId: UUID, payload: String, sessionToken: SessionToken) =
-        service.makeRequestWithPayload(
-            RequestWrapperWithPayload(POST, postDiaryStartHistoriesPath(), payload, sessionToken)
-        )
+    internal fun postDiaryStartHistory(
+        respondentId: UUID,
+        payload: String,
+        sessionToken: SessionToken
+    ) = service.makeRequestWithPayload(
+        RequestWrapperWithPayload(POST, postDiaryStartHistoriesPath(), payload, sessionToken)
+    )
 
     internal fun deleteDiaryStartHistory(respondentId: UUID, sessionToken: SessionToken) =
         service.makeRequest(RequestWrapper(DELETE, deleteDiaryStartHistoriesPath(respondentId), sessionToken))
 
 }
+
+internal fun getDiaryStartHistoriesPath(respondentId: UUID) = "/v1/diary-start-histories?respondentId=$respondentId"
+internal fun postDiaryStartHistoriesPath() = "/v1/diary-start-histories"
+internal fun deleteDiaryStartHistoriesPath(respondentId: UUID) = "/v1/diary-start-histories/respondent/$respondentId"
