+++ src/main/kotlin/no/ssb/timeusesurveyapi/questionnaire/QuestionnaireGateway.kt
@@ -39,5 +39,7 @@ class QuestionnaireGateway(
             sessionToken = sessionToken
         )
     )
-
 }
+
+internal fun getQuestionnairePath(respondentId: UUID, questionnaireType: QuestionnaireType) = "/v1/respondents/$respondentId/questionnaire/$questionnaireType"
+internal fun postQuestionnairePath(respondentId: UUID, questionnaireType: QuestionnaireType) = "/v1/respondents/$respondentId/questionnaire/$questionnaireType"
