--- src/main/kotlin/no/ssb/timeusesurveyapi/questionnaire/TimeuseSurveyServiceQuestionnairePaths.kt
@@ -1,6 +1,0 @@
-package no.ssb.timeusesurveyapi.questionnaire
-
-import java.util.*
-
-internal fun getQuestionnairePath(respondentId: UUID, questionnaireType: QuestionnaireType) = "/v1/respondents/$respondentId/questionnaire/$questionnaireType"
-internal fun postQuestionnairePath(respondentId: UUID, questionnaireType: QuestionnaireType) = "/v1/respondents/$respondentId/questionnaire/$questionnaireType"
