+++ src/main/kotlin/no/ssb/timeusesurveyapi/timeuserespondent/TimeuseRespondentGateway.kt
@@ -40,3 +40,7 @@ class TimeuseRespondentGateway(
         )
     }
 }
+
+internal fun getTimeuseRespondentByIdPath(respondentId: UUID) = "/v1/respondents/$respondentId"
+internal fun putTimeuseRespondentByIdPath(respondentId: UUID) = "/v1/respondents/$respondentId"
+internal fun putTimeuseRespondentStatusByIdPath(respondentId: UUID, status: Status, value: String) = "/v1/respondents/$respondentId/$status/$value"
